Added a player and basic interactable script
the XR Rig is inactive in this branch but still in the scene. the mouse and keyboard player can be moved around and interact with several blocks that have a basic interactable script on them to disappear when F is hit. these basic interactable objects also have an outline and prompt text when hovered over. This script can be configured with individual object scripts to make any object interactable by the pc player.